Ejercicio 3 + mejoras ejercicios anteriores

+++ practica_sistema_gestion_tareas/app/crud/task_status.py
@@ -1,5 +1,6 @@
 from sqlmodel import Session, select
 from models.task_status import TaskStatus
+from exceptions.status_task_in_use import StatusTaskInUseException
 
 def create_task_status(session: Session, task_status: TaskStatus):
     existing_task_status = session.exec(select(TaskStatus).where(TaskStatus.name == task_status.name)).first()
@@ -42,16 +43,24 @@ def update_task_status_by_name(session: Session, name: str, task_status_data: di
     return task_status
 
 def delete_task_status(session: Session, task_status_id: int):
+    from crud.task import get_tasks_by_task_status_name
     task_status = session.get(TaskStatus, task_status_id)
     if task_status:
+        in_use_status = get_tasks_by_task_status_name(session, task_status.name)
+        if in_use_status != None and len(in_use_status) > 0:
+            raise StatusTaskInUseException()
         session.delete(task_status)
         session.commit()
     return task_status
 
 def delete_task_status_by_name(session: Session, name: str):
+    from crud.task import get_tasks_by_task_status_name
     statement = select(TaskStatus).where(TaskStatus.name == name)
     task_status = session.exec(statement).first()
     if task_status:
+        in_use_status = get_tasks_by_task_status_name(session, task_status.name)
+        if in_use_status != None and len(in_use_status) > 0:
+            raise StatusTaskInUseException()
         session.delete(task_status)
         session.commit()
     return task_status
